Create How_many_unattempted_problems.py
Our Chef is currently practicing on CodeChef and is a beginner. The count of ‘All Problems’ in the Beginner section is XX. Our Chef has already ‘Attempted’ YY problems among them. How many problems are yet ‘Un-attempted’?

Input Format
The first and only line of input contains two space-separated integers XX and YY — the count of 'All problems' in the Beginner's section and the count of Chef's 'Attempted' problems, respectively.
Output Format
Output a single integer in a single line — the number of problems that are yet 'Un-attempted'

Constraints
1 \leq Y \leq X \leq 10001≤Y≤X≤1000
Subtasks
Subtask 1 (100 points):
Original constraints.