Create presents.py
Problem
Chef has fallen in love with Cheffina, and wants to buy NN gifts for her. On reaching the gift shop, Chef got to know the following two things:

The cost of each gift is 11 coin.
On the purchase of every 4^{th}4 
th
  gift, Chef gets the 5^{th}5 
th
  gift free of cost.
What is the minimum number of coins that Chef will require in order to come out of the shop carrying NN gifts?

Input Format
The first line of input will contain an integer TT — the number of test cases. The description of TT test cases follows.
The first and only line of each test case contains an integer NN, the number of gifts in the shop.
Output Format
For each test case, output on a new line the minimum number of coins that Chef will require to obtain all NN gifts.