Create ELECTIONS.py
Problem
There are 101101 citizens in Chefland. It is election time in Chefland and 33 parties, A, BA,B, and CC are contesting the elections. Party AA receives X_AX 
A
​
  votes, party BB receives X_BX 
B
​
  votes, and party CC receives X_CX 
C
​
  votes.

The constitution of Chefland requires a particular party to receive a clear majority to form the government. A party is said to have a clear majority if it receives strictly greater than 5050 votes.

If any party has a clear majority, print the winning party (A, B or C). Otherwise, print NOTA.

Input Format
The first line of input contains a single integer TT, denoting the number of test cases. The description of TT test cases follows.
The first and only line of each test case contains 33 space-separated integers — X_AX 
A
​
 , X_BX 
B
​
 , and X_CX 
C
​
 .
Output Format
For each test case, if any party has a clear majority, print the winning party (A, B or C). Otherwise, print NOTA.

You can print each letter of the string in any case (upper or lower) (for instance, strings Nota, nOtA and notA will be considered identical).